Adding to variables and printing characters, updated docs
Implemented two functions that add constants or contents of other
variables to any existing variables. Also, the printVariable function
was extended to allow printing values as their ASCII char. For this, the
least significant byte of the variable to print is extracted as a char
and added as individual item to the current session's print buffer.

Several documentation errors were fixed and the newly implemented
features documented correctly.