Implemented checks for input and output variables
The concept of input and output variables is introduced in this
change. An input variable is expected to receive input from outside the
VM in between steps, while output variables are expected to be read in
between steps. These characteristics can now be set from outside the VM,
and the program is able to determine whether any particular variable is
an input or output variable.

Tests and documentation included.