Ensured that when input is checked for new value, flag is reset
`checkIfInputWasSet` should remove the flag denoting that it was set
since it was read last. Otherwise an unnecessary read operation needs to
be performed to reset it, and with this new way, no information is
lost. The example program `feedloop` using this was corrected
accordingly, and so was the documentation.