Added new example program and refined examples in CMake file
A new example `adder` was added. It receives a stream of operands that
it should add (`+`) to each other and provide the result, which is then
compared to an expected output.

Apart from this, a new CMake switch was added that allows to explicitly
not build the examples. Also, registering examples in the CMakeLists.txt
is now done via a macro to prevent code duplication.