Implemented remaining random program factory operators
All known operators are implemented in the RandomProgramFactory class
now. Programs that are generated have a chance to be using valid ranges
of variables, string table indices, and rel/abs jump addresses. Also,
only valid operator codes are generated, and the operands applied to
them make sense in the scope of the respective operand.