Removed GALGO again
GALGO only allows to use float or double chromosomes, which makes it
unsuitable for int32_t program data. Even if one worked around
that (e.g., by scaling out 0-255 into a float) GALGO doesn't allow to
specify the initial population. This renders it absolutely unusable for
the BEAST concept with program factories that produce reasonable priors,
and multi-stage population enhancement. Removing it and looking for an
alternative.