Pipelines and GA components (#8)
* Added GALGO-2.0 3rdparty lib and some initial infra for pipelines

* Fixing GALGO repo access rules

* Added a pipe example and extended the API for pipes a bit

* Removed GALGO again

GALGO only allows to use float or double chromosomes, which makes it
unsuitable for int32_t program data. Even if one worked around
that (e.g., by scaling out 0-255 into a float) GALGO doesn't allow to
specify the initial population. This renders it absolutely unusable for
the BEAST concept with program factories that produce reasonable priors,
and multi-stage population enhancement. Removing it and looking for an
alternative.

* Adjusted Pipe's variable and function names

* Added GAlib, still need to properly configure it

* Fixed a CMake dependency to first build the galib before linking it

* Fixing a CodeQL error

* Fixed GAlib integration, now building a native CMake lib

* Hopefully fixed CodeQL and AppVeyor builds

* More AppVeyor fixes

* Even more AppVeyor fixes

* Implemented first very rough GA use-case in an example

* Removed unused item size and renamed max_size to max_candidates

* Extended the pipe example and the GA mechanisms

Still, a bug is present - the score does not improve > 0

* Fixed a bug where no finalist data would be available

* Added docs, removed debug output

* Added some Doxygen documentation

* Removed debug output and fixed code smells

* Added pipe unit tests

* Fixed compile errors