Adding pipeline applications
Pipelines consist of multiple directed evolution steps that select
candidate programs based on specific criteria. The goal of a pipeline is
to achieve a certain, usually complex, behavior. To reach this behavior,
programs need to be tailored to evolve different characteristics and
traits until they achieve the intended final behavior.

This change fixes minor issues with the VM execution code to make
evolution viable, and introduces `pipelines` example applications.