Adding EvaluatorPipe pipe class
The EvaluatorPipe pipe allows to attach arbitrary types of evaluators
to a pipe segment. This arbitrates away the need to implement pipes
for specific behaviors, and allows to implement any evaluation logic
into evaluators (where it actually belongs). This abstracts away the
raw program data from the evaluators that should only deal with
session objects.