Added a random serial data passthrough evaluator
This evaluator checks if a given program is able to forward the input
data given to it from one variable to another in a serial manner.

Also, the sorter pipeline was refactored to use this new evaluator
type in conjunction with the new EvaluatorPipe. This makes the pipeline
code much more comprehensible and the evaluator reusable.