Mutex-protected connection buffers and made connection ptrs shared
The candidate buffers in the pipeline connection objects were not
protected by mutexes. Since they are used in a multithreading context,
segmentation faults were possible. This change adds the mutex
protection. Also, it changes the previous vector implementation to a
deque one since the buffers are used as real FIFO pipes actually.

Apart from this, inside the Pipeline class, raw pointers were used for
connections in some internal places. This poses a memory leak and access
violation risk, so this was changed to use shared pointers as well.

Tests were updated accordingly.