Introduced metrics functions for pipelines
The metrics functions allow to report current performance statistics for
pipelines and individual pipes (execution count, received inputs, and
sent outputs). The reporting functions have been implemented in the
Pipeline and PipelineManager classes. Proper reporting to the web
frontend is still to be done.

Also, the handling of Pipeline class instances was changed from regular
stack objects to shared pointers to be better able to handle mutex
instances inside them. This also streamlines the overall use of shared
pointers in the software stack.